JsonParserのリファクタリング
* JSON gemの利用有無でファイルを分割
* JSON gemに習ってモジュール化
* 利用していないインスタンス変数を除去
